#99 feat: makeAssertion with argument comment
I suggest to include the argument 'comment' in assertion-functions. 
The suggestion appends the comment after the standard message and may be considered as a solution of 
issue mllg#99. The original wish was to place the comment before the standard message.
I have put the argument 'comment' before the argument 'add'. This may be a breaking change if users of the assertion functions has listed the arguments without the argument name, i.e has not used "... , add = collection", but just written (... , collection).