#99 feat: Include argument 'comment' in assert
Inclusion of the argument 'comment' in assert to make it in agrement with including the argument 'comment' in other assertion functions. In the case of 'combine = "or"', the comment is placed on a new line. In the case of 'combine = "add"', the comment is appended each assertion that fails and is on the same line. If there are more than one assertion that fails, this is not optimal.